pyproject.toml
    [build-system]
    requires = ["setuptools>=61.0", "wheel"]
    build-backend = "setuptools.build_meta"

    [project]
    name = "vitos-pizza-cafe"
    version = "1.0.0"
    description = "AI customer service application demonstrating AI security vulnerabilities and mitigation"
    readme = "README.md"
    license = {text = "MIT"}
    authors = [
        {name = "Vito's Pizza Cafe Team"}
    ]
    classifiers = [
        "Development Status :: 4 - Beta",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
    ]
    requires-python = ">=3.9"
    dependencies = [
        "streamlit>=1.44.1",
        "python-dotenv>=1.1.0",
        "langchain>=0.3.23",
        "langchain-core>=0.3.51",
        "langchain-openai>=0.3.12",
        "langchain-community>=0.3.21",
        "langgraph>=0.3.25",
        "cohere==5.14.2",
        "faiss-cpu>=1.10.0",
        "langchain-deepseek>=0.1.3",
        "langchain-cohere>=0.4.4",
        "fastapi>=0.104.1",
        "uvicorn>=0.24.0",
        "httpx>=0.25.0",
        "pydantic>=2.5.0",
        "langgraph>=0.6.7",
        "pytest>=8.4.2",
    ]

    [project.optional-dependencies]
    test = [
        "pytest>=7.4.0",
        "pytest-asyncio>=0.21.0",
        "pytest-cov>=4.1.0",
    ]
    dev = [
        "pytest>=7.4.0",
        "pytest-asyncio>=0.21.0",
        "pytest-cov>=4.1.0",
        "ruff>=0.1.0",
    ]

    [tool.setuptools.packages.find]
    where = ["."]
